# speed up forecast future